--- test/options.test.js
@@ -1,190 +1,0 @@
-const assert = require('assert')
-const babel = require('@babel/core')
-const plugin = require('../src/index.js')
-
-const React = require('react') // used in eval
-
-const ex1 = `let App = (props) => <div>{props.children}</div>; <App><div className='testclass'></div></App>`
-
-const ex2 = `let s = {}; <div className={s.testClassName}></div>`
-
-const ex3 = `<div><div /></div>`
-
-const ex4 = `let App = (props) => <div>{props.children}</div>; <App><div><span/></div></App>`
-
-const ex5 = `<><div /><div /></>`
-
-const ex6 = `let App = (props) => <div>{props.children}</div>; <App><div /><div /></App>`
-
-const ex1b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { prefix: 'testprefix', dirLevel: 0 }]]
-})
-
-const ex2b = babel.transformSync(ex2, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { addModuleClassNames: true, dirLevel: 0 }]]
-})
-
-const ex3b = babel.transformSync(ex3, {
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { ignoreTreeDepth: true }]]
-})
-
-const ex4b = babel.transformSync(ex4, {
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { firstChildOnly: true }]]
-})
-
-const ex4b2 = babel.transformSync(ex3, {
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { firstChildOnly: true }]]
-})
-
-const ex5b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { match: /fnam/ }]]
-})
-
-const ex6b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { match: /notfound/ }]]
-})
-
-const ex7b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { match: /babel/ }]] // considering git clone 'bablel-plugin-...'
-})
-
-const ex8b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { ignoreNodeNames: true, dirLevel: 0 }]]
-})
-
-const ex9b = babel.transformSync(ex1, {
-  filename: 'fname.js',
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { omitFileName: true, dirLevel: 0 }]]
-})
-
-const ex10b = babel.transformSync(ex5, {
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { firstChildOnly: true, dirLevel: 0 }]]
-})
-
-const ex11b = babel.transformSync(ex6, {
-  presets: ['@babel/preset-react'],
-  plugins: [[plugin, { firstChildOnly: true, dirLevel: 0 }]]
-})
-
-describe('Options functionality', () => {
-  describe('Prefix', function() {
-    it('should be passed', function() {
-      const output = eval(ex1b.code)
-      assert.equal(output.props['data-id'], 'testprefix__fname_App')
-    })
-  })
-
-  describe('Module Classnames', function() {
-    it('should be passed', function() {
-      const output = eval(ex2b.code)
-      assert.equal(output.props['data-id'], '_fname_div_testClassName')
-    })
-  })
-
-  describe('ignoreTreeDepth option', function() {
-    it('should ignore inner index', function() {
-      const output = eval(ex3b.code)
-      assert.equal(output.props.children.props['data-id'], '__div')
-    })
-  })
-
-  describe('firstChildOnly option', function() {
-    it('should not add data-id on component', function() {
-      const output = eval(ex4b.code)
-      assert.equal(output.props['data-id'], undefined)
-    })
-
-    it('should add data-id on first child in container', function() {
-      const output = eval(ex4b.code)
-      assert.equal(output.props.children.props['data-id'], '__div')
-    })
-
-    it('should add data-id on the only node in container', function() {
-      const output = eval(ex4b2.code)
-      assert.equal(output.props['data-id'], '__div')
-    })
-
-    it('should not add data-id on inner nodes of the only node', function() {
-      const output = eval(ex4b.code)
-      assert.equal(
-        output.props.children.props.children.props['data-id'],
-        undefined
-      )
-    })
-
-    it('should add data-id on first child in container', function() {
-      const output = eval(ex11b.code)
-      assert.equal(output.props.children[0].props['data-id'], '__div')
-    })
-
-    it('should not add data-id on the second child in container', function() {
-      const output = eval(ex11b.code)
-      assert.equal(output.props.children[1].props['data-id'], undefined)
-    })
-
-    it('should add data-id on first child of a fragment', function() {
-      const output = eval(ex10b.code)
-      assert.equal(output.props.children[0].props['data-id'], '__div')
-    })
-
-    it('should not add data-id on second child of a fragment', function() {
-      const output = eval(ex10b.code)
-      assert.equal(output.props.children[1].props['data-id'], undefined)
-    })
-  })
-
-  describe('match option', function() {
-    it('should add data-id on matching component (filename)', function() {
-      const output = eval(ex5b.code)
-      assert.notEqual(output.props['data-id'], undefined)
-      assert.notEqual(output.props.children.props['data-id'], undefined)
-    })
-
-    it('should add data-id on matching component (dirname)', function() {
-      const output = eval(ex7b.code)
-      assert.notEqual(output.props['data-id'], undefined)
-      assert.notEqual(output.props.children.props['data-id'], undefined)
-    })
-
-    it('should add correct regex prefix', function() {
-      const output = eval(ex5b.code)
-      assert.equal(output.props['data-id'].startsWith('fnam_'), true)
-    })
-
-    it('should not add data-id on non-matching component', function() {
-      const output = eval(ex6b.code)
-      assert.equal(output.props['data-id'], undefined)
-    })
-  })
-
-  describe('ignoreNodeNames option', function() {
-    it('should ignore node name', function() {
-      const output = eval(ex8b.code)
-      assert.equal(output.props.children.props['data-id'], '_fname')
-    })
-  })
-
-  describe('omitFileName option', function() {
-    it('should ignore node name', function() {
-      const output = eval(ex9b.code)
-      assert.equal(output.props.children.props['data-id'], 'div')
-    })
-  })
-})
